pyproject.toml
[project]
name = "introduction-to-lang-chain"
version = "0.1.0"
description = ""
authors = [
    {name = "Jarryd Deane",email = "jarryd.deane@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-ollama (>=0.3.6,<0.4.0)",
    "langchain-anthropic (>=0.3.19,<0.4.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "text-generation (>=0.7.0,<0.8.0)",
    "transformers (>=4.55.3,<5.0.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "numexpr (>=2.11.0,<3.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "bitsandbytes (>=0.47.0,<0.48.0)",
    "accelerate (>=1.10.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
